requirements.txt
pygithub
requests
click
pyyaml
matplotlib
android-category>=0.0.4
